refactor(studio-ui): split prompt input primitives
